third_party/blink/web_tests/http/tests/devtools/tracing/timeline-layout/timeline-layout-with-invalidations-expected.txt

Tests the Timeline API instrumentation of layout events with invalidations.


Running: testLocalFrame
first layout invalidations[
    {
        cause : {reason: Style changed, stackTrace: test://evaluations/0/timeline-layout-with-invalidations.js:23}
        changedAttribute : undefined
        changedClass : undefined
        changedId : undefined
        changedPseudo : undefined
        extraData : undefined
        nodeName : "DIV id='testElement'"
        selectorPart : undefined
        type : "LayoutInvalidationTracking"
    }
]
second layout invalidations[
    {
        cause : {reason: Style changed, stackTrace: test://evaluations/0/timeline-layout-with-invalidations.js:25}
        changedAttribute : undefined
        changedClass : undefined
        changedId : undefined
        changedPseudo : undefined
        extraData : undefined
        nodeName : "DIV id='testElement'"
        selectorPart : undefined
        type : "LayoutInvalidationTracking"
    }
]

Running: testSubframe
first layout invalidations[
    {
        cause : {reason: Style changed, stackTrace: test://evaluations/0/timeline-layout-with-invalidations.js:32}
        changedAttribute : undefined
        changedClass : undefined
        changedId : undefined
        changedPseudo : undefined
        extraData : undefined
        nodeName : "DIV"
        selectorPart : undefined
        type : "LayoutInvalidationTracking"
    }
]
second layout invalidations[
    {
        cause : {reason: Style changed, stackTrace: test://evaluations/0/timeline-layout-with-invalidations.js:34}
        changedAttribute : undefined
        changedClass : undefined
        changedId : undefined
        changedPseudo : undefined
        extraData : undefined
        nodeName : "DIV"
        selectorPart : undefined
        type : "LayoutInvalidationTracking"
    }
]
